price changed from big decimal to double, some unused methods removed

+++ src/main/java/Product.java
@@ -1,7 +1,7 @@
-import java.math.BigDecimal;
+//import java.math.BigDecimal;
 
 
-public record Product(String nameOfProduct, BigDecimal priceOfProduct, int idOfProduct) {
+public record Product(String nameOfProduct, double priceOfProduct, int idOfProduct) {
